styles shifted to styles folder
