+++ components/profile/ProfileDropdown.tsx
@@ -1,11 +1,12 @@
 import { useState } from "react";
-import { Button, StyleSheet } from "react-native";
+import { Button } from "react-native";
 import { useRouter } from "expo-router";
 
 import { View } from "../Themed";
 import { useAuthContext } from "@/contexts";
 import ProfilePopup from "./ProfilePopup";
 import { toast } from "react-toastify";
+import styles from "@/styles/profile/Profile.styles";
 
 const ProfileDropdown = () => {
   const [isProfilePopupVisible, setProfilePopupVisible] = useState(false);
@@ -42,27 +43,5 @@ const ProfileDropdown = () => {
     </>
   );
 };
-
-const styles = StyleSheet.create({
-  dropdown: {
-    position: 'absolute',
-    top: '100%',
-    right: 0,
-    backgroundColor: '#fff',
-    padding: 10,
-    borderRadius: 5,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.2,
-    shadowRadius: 2,
-    elevation: 5,
-    marginTop: 10,
-    gap: 10,
-    width: 140,
-  },
-  dropdownText: {
-    fontSize: 16,
-  },
-});
 
 export default ProfileDropdown;
