+++ components/profile/ProfilePopup.tsx
@@ -1,5 +1,6 @@
-import { Button, Modal, StyleSheet } from "react-native";
+import { Button, Modal } from "react-native";
 import { Text, View } from "../Themed";
+import styles from "@/styles/profile/Profile.styles";
 
 interface ProfilePopupProps {
   visible: boolean;
@@ -16,7 +17,7 @@ const ProfilePopup: React.FC<ProfilePopupProps> = ({
       transparent={true}
       visible={visible}
     >
-      <View style={styles.container}>
+      <View style={styles.profilePopupContainer}>
         <View style={styles.popup}>
           <Text style={styles.title}>Edit Profile</Text>
           <Button title="Close" onPress={() => onClose()} />
@@ -25,26 +26,5 @@ const ProfilePopup: React.FC<ProfilePopupProps> = ({
     </Modal>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-  },
-  popup: {
-    width: '80%',
-    padding: 20,
-    backgroundColor: 'white',
-    borderRadius: 10,
-    alignItems: 'center',
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    marginBottom: 20,
-  },
-});
 
 export default ProfilePopup;
