+++ styles/header/Header.styles.tsx
@@ -1,0 +1,17 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  headerContainer: {
+    alignItems: 'center',
+    borderBottomColor: 'lightgray',
+    borderBottomWidth: 1,
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    zIndex: 1000,
+  },
+  headerSection: {
+    flexDirection: 'row',
+  }
+});
+
+export default styles;
