+++ styles/profile/Profile.styles.tsx
@@ -1,0 +1,49 @@
+import Colors from "@/constants/Colors";
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    position: 'relative',
+  },
+  profileIconContainer: {
+    alignItems: 'center',
+    backgroundColor: Colors.regular.primary,
+    justifyContent: 'center',
+  },
+  dropdown: {
+    position: 'absolute',
+    top: '100%',
+    right: 0,
+    backgroundColor: '#fff',
+    padding: 10,
+    borderRadius: 5,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.2,
+    shadowRadius: 2,
+    elevation: 5,
+    marginTop: 10,
+    gap: 10,
+    width: 140,
+  },
+  profilePopupContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: 'rgba(0, 0, 0, 0.5)',
+  },
+  popup: {
+    width: '80%',
+    padding: 20,
+    backgroundColor: 'white',
+    borderRadius: 10,
+    alignItems: 'center',
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    marginBottom: 20,
+  },
+});
+
+export default styles;
